feat: Add scroll-to-bottom button in ChatComponent
This commit adds a scroll-to-bottom button in the ChatComponent to allow users to easily navigate to the latest messages. The button is displayed when the user scrolls more than 300 pixels from the top of the page. Clicking the button triggers a smooth scroll animation to the bottom of the chat. This enhancement improves the user experience by providing a convenient way to view new messages without manually scrolling.